Don't read environment variables without prefix
If envconfig can't find an environment variable in the form PREFIX_MYVAR, and
there is a struct tag defined, it will try to populate your variable with an
environment variable that directly matches the envconfig tag in your struct
definition. To avoid this behavior we are removing the envconfig tag from the
configuration structure.