Fix synchronization problem on backup retrieval
There're some actions performed locally that aren't synchronized with
the cloud. For example, when removing a backup we replace the file
references of the removed backup in other backups, and many archive
information in the cloud gets outdated. So we will check if all
references in the remote archive information are valid before replacing
the local version.

We also fixed how we detect the backup parts that we should download, we should
only ignore deleted backup parts.